Updated benchmark, moved PathSelect to JsonPathSelect.  Updated read.me with benchmark data
